Small refactor of enterprise class and loading sequence.
