+++ Frontend/app/shared/enterprise.ts
